+++ texproject/template.py
@@ -1,21 +1,23 @@
 from jinja2 import Environment, FileSystemLoader
 import datetime
 from pathlib import Path
+import os
+import errno
 
 from .filesystem import (DATA_DIR, TPR_INFO_FILENAME, CONFIG,
         _TEMPLATE_DOC_NAME, _PROJECT_MACRO_TEMPLATE,
-        _CLASSINFO_TEMPLATE,_BIBINFO_TEMPLATE,
+        _CLASSINFO_TEMPLATE,_BIBINFO_TEMPLATE,_BIBLIOGRAPHY_TEMPLATE,
         TEMPLATE_RESOURCE_DIR, load_user_dict, yaml_dump_proj_info, load_proj_dict,
-        macro_loader, format_loader, citation_loader, template_loader)
+        macro_linker, format_linker, citation_linker, template_linker)
 
 # special renamed function for use in templates
 def safe_name(name, style):
     if style == 'macro':
-        return macro_loader.safe_name(name)
+        return macro_linker.safe_name(name)
     elif style == 'citation':
-        return citation_loader.safe_name(name)
+        return citation_linker.safe_name(name)
     elif style == 'format':
-        return format_loader.safe_name(name)
+        return format_linker.safe_name(name)
     else:
         return name
 
@@ -40,6 +42,9 @@ class GenericTemplate:
                 )
 
         self.env.filters['safe_name'] = safe_name
+        # bootstrap bibliography content
+        self.bibtext = self.env.get_template(
+                str(_BIBLIOGRAPHY_TEMPLATE)).render(config = CONFIG)
 
 
     def render_template(self, template):
@@ -47,22 +52,25 @@ class GenericTemplate:
             user = self.user_dict, # user parameters
             template = self.template_dict, # template parameters
             config = CONFIG, # general filename conventions
-            bibliography = f"\\input{{{CONFIG['bibinfo_file']}}}",
+            bibliography = self.bibtext,
             date=datetime.date.today())
 
 
-    def write_template(self, template_path, target_path):
-        # warn if file already exists??
-        return target_path.write_text(
-                self.render_template(
-                    self.env.get_template(str(template_path))))
+    def write_template(self, template_path, target_path, force=False):
+        if target_path.exists() and not force:
+            raise FileExistsError(
+                    errno.EEXIST,
+                    f"Template write location aready exists",
+                    str(target_path.resolve()))
+        target_path.write_text(
+            self.render_template(
+                self.env.get_template(str(template_path))))
 
 class ProjectTemplate(GenericTemplate):
     @classmethod
-    def load_from_template(cls, template_name, project_name, citations,frozen=False):
-        template_dict = template_loader.load_template(template_name)
+    def load_from_template(cls, template_name, citations, frozen=False):
+        template_dict = template_linker.load_template(template_name)
         template_dict['citations'].extend(citations)
-        template_dict['project'] = project_name
         template_dict['frozen'] = frozen
         self = cls(template_dict)
         self.template_path = Path('templates', template_name, _TEMPLATE_DOC_NAME)
@@ -74,39 +82,45 @@ class ProjectTemplate(GenericTemplate):
         return cls(template_dict)
 
 
-    def write_tpr_files(self, out_folder,force=False):
+    def write_tpr_files(self, out_folder,force=False,write_template=False):
+        project_folder = out_folder / CONFIG['project_folder']
+        project_folder.mkdir(exist_ok=True)
+        # write project information file
+        if write_template:
+            yaml_dump_proj_info(out_folder, self.template_dict)
+
         # write templates
         self.write_template(
                 _CLASSINFO_TEMPLATE,
-                out_folder / f"{CONFIG['classinfo_file']}.tex")
+                project_folder / f"{CONFIG['classinfo_file']}.tex",
+                force=True)
         self.write_template(
                 _BIBINFO_TEMPLATE,
-                out_folder / f"{CONFIG['bibinfo_file']}.tex")
+                project_folder / f"{CONFIG['bibinfo_file']}.tex",
+                force=True)
 
         # link macro, format, and citation files from resources
         for macro in self.template_dict['macros']:
-            macro_loader.link_name(macro, out_folder,
-                    frozen=self.template_dict['frozen'],force=force)
+            macro_linker.link_name(macro, project_folder,
+                    frozen=self.template_dict['frozen'],
+                    force=force)
         for cit in self.template_dict['citations']:
-            citation_loader.link_name(cit, out_folder,
+            citation_linker.link_name(cit, project_folder,
                     frozen=self.template_dict['frozen'],force=force)
-        format_loader.link_name(
+        format_linker.link_name(
                 self.template_dict['format'],
-                out_folder,
+                project_folder,
                 frozen=self.template_dict['frozen'],force=force)
 
     def create_output_folder(self, out_folder):
-        out_folder.mkdir(parents=True,exist_ok=True)
+        project_folder = out_folder / CONFIG['project_folder']
 
         # write local files from templates
         self.write_template(
                 self.template_path,
-                out_folder / f"{self.template_dict['project']}.tex")
+                out_folder / f"{CONFIG['default_tex_name']}.tex")
         self.write_template(
                 _PROJECT_MACRO_TEMPLATE,
                 out_folder / f"{CONFIG['project_macro_file']}.sty")
 
-        # write project information file
-        yaml_dump_proj_info(out_folder, self.template_dict)
-
-        self.write_tpr_files(out_folder)
+        self.write_tpr_files(out_folder,write_template=True)
